requirements.txt
torch
torchvision
albumentations
opencv-python-headless
numpy
matplotlib
scikit-image
tqdm
Pillow
scipy
